Refs #89. Added DialogBox class.
DialogBoxes can be transient or persistent.

A persistent dialog box is stored into the User session data,
waiting for the User to dismiss it, by clicking on one of its buttons
(actions).

A transient dialog box does not have any button and is displayed until is dismissed or the interface is refreshed.

Dialog boxes are managed by NotificationArea module.